+++ src/main/java/com/mauriciotogneri/idlebattle/messages/MatchStatus.java
@@ -5,10 +5,10 @@ import java.util.List;
 public class MatchStatus
 {
     private final String id;
-    private final double remainingTime;
+    private final int remainingTime;
     private final List<LaneStatus> lanes;
 
-    public MatchStatus(String id, double remainingTime, List<LaneStatus> lanes)
+    public MatchStatus(String id, int remainingTime, List<LaneStatus> lanes)
     {
         this.id = id;
         this.remainingTime = remainingTime;
